refactor: retire root storage compat facade
