+++ docs/architecture/migration-progress.md
@@ -6,15 +6,16 @@ Status values: `[ ]` not started, `[~]` in progress, `[x]` complete, `[!]` block
 
 - Issue: [`rustfs/backlog#660`](https://github.com/rustfs/backlog/issues/660)
 - Branch: `overtrue/arch-runtime-compat-consumer-batch`
-- Baseline: completed `C-011/C-012/C-013/API-055/API-059/API-079/API-080/API-081/API-082/API-083/API-084/API-085/API-086/API-087/API-088/API-089/API-090/API-091/API-092/API-093/API-094/API-095/API-096/API-097/API-098/API-099`.
+- Baseline: completed `C-011/C-012/C-013/API-055/API-059/API-079/API-080/API-081/API-082/API-083/API-084/API-085/API-086/API-087/API-088/API-089/API-090/API-091/API-092/API-093/API-094/API-095/API-096/API-097/API-098/API-099/API-100`.
 - Stacked on: `overtrue/arch-disk-rpc-method-wrappers` pending API-097 merge.
 - PR type for this branch: `pure-move`
 - Runtime behavior changes: none.
-- Rust code changes: move capacity, server, and startup ECStore compatibility
-  consumers from the root runtime facade into local boundaries.
-- CI/script changes: guard against restoring capacity/server/startup consumer wrappers
-  in the root runtime compatibility facade.
-- Docs changes: record the API-098/API-099 runtime consumer boundary cleanup.
+- Rust code changes: move capacity, server, startup, table catalog,
+  runtime-capability, workload-admission, error, and config-test ECStore
+  compatibility consumers from the root runtime facade into local boundaries.
+- CI/script changes: guard against restoring consumer wrappers in the root
+  runtime compatibility facade and tolerate retiring the empty root facade.
+- Docs changes: record the API-098/API-099/API-100 runtime consumer boundary cleanup.
 
 ## Phase 0 Tasks
 
@@ -419,6 +420,21 @@ Status values: `[ ]` not started, `[~]` in progress, `[x]` complete, `[!]` block
   - Verification: RustFS test-target compile coverage, startup residual scan,
     migration and layer guards, formatting, diff hygiene, Rust risk scan,
     pre-commit quality gate, and three-expert review.
+- [x] `API-100` Retire root runtime storage compatibility consumers.
+  - Completed slice: move table catalog metadata constants and bucket metadata
+    reads, runtime topology capability mapping, workload admission runtime
+    state probes, S3 error mapping aliases, and config test disk-layout aliases
+    into local compatibility boundaries, then remove the root
+    `storage_compat.rs` module.
+  - Acceptance: no RustFS source consumes `crate::storage_compat`; root
+    runtime compatibility file is removed; migration rules still reject direct
+    ECStore imports outside `*storage_compat.rs` boundaries.
+  - Must preserve: table catalog internal metadata paths, lock timeout lookup,
+    runtime topology snapshots, workload admission status reporting, quota and
+    storage error mapping, and config disk-layout parsing tests.
+  - Verification: RustFS test-target compile coverage, direct root compatibility
+    consumer residual scan, migration and layer guards, formatting, diff
+    hygiene, Rust risk scan, pre-commit quality gate, and three-expert review.
 - [x] `G-012` Inventory placement and repair invariants.
   - Acceptance:
     [`placement-repair-invariants.md`](placement-repair-invariants.md) records
@@ -3482,6 +3498,18 @@ Passed before push:
   - `./scripts/check_layer_dependencies.sh`: passed.
   - Startup/init root compatibility consumer residual scan: passed.
   - Root startup consumer wrapper residual scan: passed.
+  - Rust risk scan on changed Rust files and guard script: passed.
+  - `make pre-commit`: passed.
+
+- Issue #660 API-100 current slice:
+  - `cargo check -p rustfs --tests`: passed.
+  - `cargo fmt --all`: passed.
+  - `cargo fmt --all --check`: passed.
+  - `git diff --check`: passed.
+  - `bash -n scripts/check_architecture_migration_rules.sh`: passed.
+  - `./scripts/check_architecture_migration_rules.sh`: passed.
+  - `./scripts/check_layer_dependencies.sh`: passed.
+  - Direct root compatibility consumer residual scan: passed.
   - Rust risk scan on changed Rust files and guard script: passed.
   - `make pre-commit`: passed.
